pyproject.toml
[tool.poetry]
name = "djangomongosrl"
version = "0.1.0"
description = ""
authors = ["gwenf <gwenfaraday@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.7.9"
djangorestframework = "^3.12.2"
djongo = "^1.3.3"
Django = "3.0.5"
drf-nested-routers = "^0.92.5"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
